refactor: 도메인별 Finder 클래스 분리로 중복된 조회 로직 제거 - User, Place 등의 도메인 객체 조회 시마다 발생하던 repository 호출 및 예외 처리 중복 제거 - 각 도메인에 Finder 클래스를 생성해 객체 탐색의 책임을 위임하고, 서비스 로직의 응집도 향상
