+++ src/main/java/com/triplog/Bookmark/repository/BookmarkRepository.java
@@ -1,11 +1,10 @@
-package com.triplog.Bookmark;
+package com.triplog.Bookmark.repository;
 
 import com.triplog.Bookmark.domain.Bookmark;
 import com.triplog.place.domain.Place;
 import com.triplog.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.awt.print.Book;
 import java.util.Optional;
 
 public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
