+++ src/main/java/com/triplog/place/PlaceFinder.java
@@ -1,0 +1,20 @@
+package com.triplog.place;
+
+import com.triplog.common.exception.CustomException;
+import com.triplog.common.exception.ErrorCode;
+import com.triplog.place.domain.Place;
+import com.triplog.place.repository.PlaceRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class PlaceFinder {
+
+    private final PlaceRepository placeRepository;
+
+    public Place findByKakaoPlaceId(Long kakaoPlaceId) {
+        return placeRepository.findByKakaoPlaceId(kakaoPlaceId)
+                .orElseThrow(() -> new CustomException(ErrorCode.PLACE_NOT_FOUND));
+    }
+}
