+++ src/main/java/com/triplog/place/controller/PlaceController.java
@@ -1,10 +1,10 @@
-package com.triplog.place;
+package com.triplog.place.controller;
+import com.triplog.place.service.PlaceService;
 import com.triplog.place.dto.PlaceSaveRequest;
 import com.triplog.place.dto.PlaceSaveResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
