+++ src/main/java/com/triplog/user/UserFinder.java
@@ -1,0 +1,20 @@
+package com.triplog.user;
+
+import com.triplog.common.exception.CustomException;
+import com.triplog.common.exception.ErrorCode;
+import com.triplog.user.domain.User;
+import com.triplog.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class UserFinder {
+
+    private final UserRepository userRepository;
+
+    public User findByNickname(String nickname) {
+        return userRepository.findByNickname(nickname)
+                .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+    }
+}
